# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Gonemaster API
  version: 0.1.0
  description: |
    REST API for running Gonemaster engine tests with queueing, streaming, and cancellation.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development (API base)
...
paths:
...
  /session:
    post:
      tags: [Auth]
      summary: Admin UI login - validate a token and set the session cookie
      operationId: sessionLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        '200':
          description: Authenticated; sets the gm_admin session cookie
        '401':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Auth]
      summary: Admin UI logout - clear the session cookie
      operationId: sessionLogout
      responses:
        '200':
          description: Session cookie cleared
  /jobs:
    post:
      tags: [Jobs]
      summary: Submit a single test job
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/ErrorResponse'
    get:
      tags: [Jobs]
      summary: List jobs
      operationId: listJobs
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
        - name: batch_id
          in: query
          schema:
            type: string
        - name: created_after
          in: query
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          schema:
            type: string
            format: date-time
        - name: domain
          in: query
          schema:
            type: string
        - name: severity
          in: query
          description: Filter by aggregated severity totals before pagination.
          schema:
            type: string
            enum: [warnings_plus, errors_only]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Job list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobList'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /jobs/batch:
    post:
      tags: [Jobs]
      summary: Submit a batch of test jobs
      description: "Submit multiple domains as a single batch. Each job is assigned priority=1 (batch) and will only be dequeued after all priority=0 (normal) jobs have been processed. Undelegated input (`nameservers`, `ds_info`) is not supported for batch submission."
      operationId: createJobBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobBatchRequest'
      responses:
        '202':
          description: Batch accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobBatchResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /jobs/purge:
    post:
      tags: [Jobs]
      summary: Purge completed jobs older than a retention window
      operationId: purgeJobs
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobPurgeRequest'
      responses:
        '200':
          description: Purge count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobPurgeResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /batches/{batch_id}:
    get:
      tags: [Batches]
      summary: Get batch summary
      operationId: getBatch
      parameters:
        - name: batch_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSummary'
        '404':
          $ref: '#/components/responses/ErrorResponse'
    delete:
      tags: [Batches]
      summary: Delete a batch and derived records
      operationId: deleteBatch
      parameters:
        - name: batch_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Batch deleted
        '404':
          $ref: '#/components/responses/ErrorResponse'
        '409':
          $ref: '#/components/responses/ErrorResponse'
  /batches/{batch_id}/delete-preview:
    get:
      tags: [Batches]
      summary: Preview batch deletion impact
      operationId: getBatchDeletePreview
      parameters:
        - name: batch_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Batch deletion impact
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchDeletePreview'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /jobs/{job_id}:
    get:
      tags: [Jobs]
      summary: Get job status
      operationId: getJob
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /jobs/{job_id}/result:
    get:
      tags: [Jobs]
      summary: Get job result
      operationId: getJobResult
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: locale
          in: query
          description: Locale for translated messages (default en)
          schema:
            type: string
      responses:
        '200':
          description: Job result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResult'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
  /domains:
    get:
      tags: [Domains]
      summary: List domains
      operationId: listDomains
      parameters:
        - name: tag
          in: query
          schema:
            type: string
          description: Filter to domains belonging to this tag.
        - name: name
          in: query
          schema:
            type: string
          description: Substring filter on domain name.
        - name: level
          in: query
          schema:
            type: string
          description: Exact `latest_level` filter (e.g. `ERROR`).
        - name: min_level
          in: query
          schema:
            type: string
          description: Minimum severity threshold; `WARNING` matches WARNING/ERROR/CRITICAL.
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Domain list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DomainList'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /domains/{id}:
    get:
      tags: [Domains]
      summary: Get domain by ID
      operationId: getDomain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Domain with tags populated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Domain'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /domains/{id}/runs:
    get:
      tags: [Domains]
      summary: List run history for a domain
      operationId: listDomainRuns
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Run list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /runs:
    get:
      tags: [Runs]
      summary: List runs
      operationId: listRuns
      parameters:
        - name: tag
          in: query
          schema:
            type: string
          description: Filter to runs for domains in this tag.
        - name: domain
          in: query
          schema:
            type: string
          description: Filter by domain name substring.
        - name: batch
          in: query
          schema:
            type: string
          description: Filter by batch ID.
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/JobStatus'
        - name: level
          in: query
          schema:
            type: string
          description: Filter by `worst_level`.
        - name: grade
          in: query
          schema:
            type: string
            enum: ['A+', 'A', 'B', 'C', 'D', 'F']
          description: Filter by letter grade.
        - name: finished_after
          in: query
          schema:
            type: string
            format: date-time
        - name: finished_before
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Run list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunList'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /runs/{id}:
    get:
      tags: [Runs]
      summary: Get run by ID
      operationId: getRun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /runs/{id}/result:
    get:
      tags: [Runs]
      summary: Get run result (same shape as GET /jobs/{id}/result)
      operationId: getRunResult
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: locale
          in: query
          description: Locale for translated messages (default en)
          schema:
            type: string
      responses:
        '200':
          description: Run result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResult'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /runs/{id}/dnssec-chain:
    get:
      tags: [Runs]
      summary: Get the stored DNSSEC chain summary for a run
      description: >
        Returns the per-run DNSSEC authentication-chain summary extracted at
        run end. Only runs created through the public UI ever have chain data.
      operationId: getRunDNSSECChain
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: DNSSEC chain summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DNSSECChain'
        '404':
          $ref: '#/components/responses/ErrorResponse'
  /spec/testcases:
    get:
      tags: [Spec]
      summary: List implemented testcases
      operationId: listSpecTestcases
      parameters:
        - name: category
          in: query
          description: "Filter to one module (alias: module)"
          schema:
            type: string
      responses:
        '200':
          description: Testcase catalog
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecTestcaseList'
...
  /tags/{name}/summary:
    get:
      tags: [Tags]
      summary: Get tag severity summary
      operationId: getTagSummary
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Per-severity domain counts for the tag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagSummary'
        '404':
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >-
        Admin token. Required on /api/v1 endpoints when the server is started
        with admin tokens configured (token mode); ignored in open mode. Clients
        send "Authorization: Bearer <token>"; the admin UI uses a session cookie
        set via POST /session.
...
  schemas:
...
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, canceled, expired, paused]
...
    JobBatchResponse:
      type: object
      required: [batch_id, job_ids]
      properties:
        batch_id:
          type: string
        job_ids:
          type: array
          items:
            type: string
    JobPurgeRequest:
      type: object
      properties:
        older_than_days:
          type: integer
          minimum: 0
          description: Days before now. Zero uses the server retention setting.
    JobPurgeResponse:
      type: object
      required: [purged_jobs]
      properties:
        purged_jobs:
          type: integer
          format: int64
...
    JobList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        total:
          type: integer
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'